refactor(metrics): split GTFS-RT vehicle logic into store and metrics files
- move VehicleLastSeen struct & methods to vehicle_store.go
- keep metric calculation functions in vehicle_metrics.go
- rename gtfs-realtime-bindings_test.go to vehicle_metrics_test.go
- improve separation of concerns between state management and business logic
