+++ internal/metrics/vehicle_metrics_test.go
